+++ src/include/crf/state.h
@@ -10,13 +10,13 @@ namespace NLP {
 
         State(const size_t ntags)
           : words(), pos(), entities(), lattice(ntags), dist() {
-          for (int i = 0; i < ntags; ++i)
+          for (size_t i = 0; i < ntags; ++i)
             dist.push_back(PDF(ntags, 0.0));
         }
 
         void reset(void) {
           lattice.reset();
-          for (int i = 0; i < dist.size(); ++i)
+          for (size_t i = 0; i < dist.size(); ++i)
             std::fill(dist[i].begin(), dist[i].end(), 0.0);
         }
     };
